# docs/swagger.yaml
basePath: /api/
definitions:
...
  domain.CommunityForMyCommunity:
    properties:
      avatarPath:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  domain.CommunityForView:
    properties:
      avatarPath:
        type: string
      coverPath:
        type: string
      createdAt:
        type: string
      creatorID:
        type: integer
      description:
        type: string
      id:
        type: integer
      isSubscribed:
        type: boolean
      name:
        type: string
      subscribersCount:
        type: integer
    type: object
  domain.CommunitySubscriber:
    properties:
      avatarPath:
        type: string
      fullName:
        type: string
      userID:
        type: integer
    type: object
...
  domain.ShortCommunity:
    properties:
      avatarPath:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      subscribersCount:
        type: integer
    type: object
...
  handler.JSONResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a VK API.
  termsOfService: http://swagger.io/terms/
  title: VK API
  version: "1.0"
paths:
...
  /communities:
    post:
      consumes:
      - multipart/form-data
      description: Создает новое сообщество с возможностью загрузки аватара и обложки
      parameters:
      - description: Название сообщества (3-48 символов)
        in: formData
        name: name
        required: true
        type: string
      - description: Описание сообщества (до 512 символов)
        in: formData
        name: description
        type: string
      - description: Аватар сообщества
        in: formData
        name: avatar
        type: file
      - description: Обложка сообщества
        in: formData
        name: cover
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Сообщество успешно создано
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать сообщество
      tags:
      - communities
  /communities/{id}:
    delete:
      description: Удаляет сообщество (только создатель)
      parameters:
      - description: ID сообщества
        format: int32
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Сообщество успешно удалено
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверный ID сообщества
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "403":
          description: Доступ запрещен (не создатель)
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Сообщество не найдено
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить сообщество
      tags:
      - communities
    get:
      description: Возвращает информацию о сообществе включая количество подписчиков,
        статус подписки текущего пользователя, создателя
      parameters:
      - description: ID сообщества
        format: int32
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о сообществе
          schema:
            $ref: '#/definitions/domain.CommunityForView'
        "400":
          description: Неверный ID сообщества
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Сообщество не найдено
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить информацию о сообществе
      tags:
      - communities
    put:
      consumes:
      - multipart/form-data
      description: Обновляет информацию о сообществе (только создатель)
      parameters:
      - description: ID сообщества
        format: int32
        in: path
        name: id
        required: true
        type: integer
      - description: Название сообщества (3-48 символов)
        in: formData
        name: name
        type: string
      - description: Описание сообщества (до 512 символов)
        in: formData
        name: description
        type: string
      - description: Новый аватар сообщества
        in: formData
        name: avatar
        type: file
      - description: Новая обложка сообщества
        in: formData
        name: cover
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Сообщество успешно обновлено
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "403":
          description: Доступ запрещен (не создатель)
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Сообщество не найдено
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить сообщество
      tags:
      - communities
  /communities/{id}/subscribe:
    post:
      description: Подписывает текущего пользователя на указанное сообщество
      parameters:
      - description: ID сообщества
        format: int32
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешная подписка
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверный ID сообщества
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Сообщество не найдено
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Подписаться на сообщество
      tags:
      - communities
  /communities/{id}/subscribers:
    get:
      description: Возвращает список подписчиков указанного сообщества
      parameters:
      - description: ID сообщества
        format: int32
        in: path
        name: id
        required: true
        type: integer
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество подписчиков на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список подписчиков
          schema:
            items:
              $ref: '#/definitions/domain.CommunitySubscriber'
            type: array
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Сообщество не найдено
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить подписчиков сообщества
      tags:
      - communities
  /communities/{id}/unsubscribe:
    post:
      description: Отписывает текущего пользователя от указанного сообщества
      parameters:
      - description: ID сообщества
        format: int32
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешная отписка
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "400":
          description: Неверный ID сообщества
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "404":
          description: Подписка не найдена
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Отписаться от сообщества
      tags:
      - communities
  /communities/created:
    get:
      description: Возвращает список сообществ, созданных текущим пользователем (только
        ID, название и аватар)
      parameters:
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество сообществ на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список созданных сообществ
          schema:
            items:
              $ref: '#/definitions/domain.CommunityForMyCommunity'
            type: array
        "400":
          description: Неверные параметры пагинации
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить созданные сообщества
      tags:
      - communities
  /communities/my:
    get:
      description: Возвращает список сообществ, на которые подписан пользователь
      parameters:
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество сообществ на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список сообществ пользователя
          schema:
            items:
              $ref: '#/definitions/domain.ShortCommunity'
            type: array
        "400":
          description: Неверные параметры пагинации
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить сообщества пользователя
      tags:
      - communities
  /communities/other:
    get:
      description: Возвращает список сообществ, на которые пользователь не подписан
        (рекомендации)
      parameters:
      - default: 1
        description: Номер страницы
        format: int32
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: Количество сообществ на странице
        format: int32
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список рекомендуемых сообществ
          schema:
            items:
              $ref: '#/definitions/domain.ShortCommunity'
            type: array
        "400":
          description: Неверные параметры пагинации
          schema:
            $ref: '#/definitions/handler.JSONResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.JSONResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить другие сообщества
      tags:
      - communities
...
swagger: "2.0"